Type, header refactoring; simple random numbers

+++ libEQ/Eq.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int Eq(long double a, long double b, long double e);
+int Eq(double a, double b, double e);
